requirements.txt
PyQt5==5.15.11
sip==6.8.6
Pillow==9.5.0
requests==2.31.0
openai==0.27.0
setuptools==75.1.0
mss==9.0.2